requirements.txt
Pylons==1.0.2
waitress==0.8.9
Paste==2.0.2
PasteDeploy==1.5.2
PasteScript==2.0.2
PyYAML==3.11